Fixing generics adjMatrix grapha

+++ src/main/java/graphEstructure/GraphInterface.java
@@ -9,7 +9,7 @@ public interface GraphInterface<T> {
 	
 	public void addVertex(T element);
 	public void addEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
-	public void addEdge(Vertex<T> from, Vertex<T> to, double weight)throws IllegalArgumentException;
+	public void addEdge(Vertex<T> from, Vertex<T> to, int weight)throws IllegalArgumentException;
 	public void removeVertex(Vertex<T> vertex);
 	public void removeEdge(Vertex<T> from, Vertex<T> to)throws IllegalArgumentException;
 	public Vertex<T> getVertex(T element) throws IllegalArgumentException;
@@ -17,7 +17,7 @@ public interface GraphInterface<T> {
 	public Edge<T> getEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
 	public Iterable<Edge<T>> getEdges(Vertex<T> vertex) throws IllegalArgumentException;
 	public double getWeightEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
-	public void setWeightEdge(Vertex<T> from, Vertex<T> to, double weight) throws IllegalArgumentException;
+	public void setWeightEdge(Vertex<T> from, Vertex<T> to, int weight) throws IllegalArgumentException;
 	public boolean isAdjacent(Vertex<T> from, Vertex<T> to);
 	public Set<Vertex<T>> adjacentVertices(Vertex<T> vertex);
 	public Iterable<Edge<T>> getEdges();
